Create edit delete bugs fixed in expense screen

+++ src/app/shared/data.service.ts
@@ -17,6 +17,7 @@ export class DataService {
   ) {}
 
   addExpense(expense: Expense) {
+    console.log('Add Expense Triggred');
     return this.fireauth.currentUser.then((user) => {
       if (user) {
         expense.id = this.afs.createId();
@@ -35,7 +36,7 @@ export class DataService {
             .collection(`/Users/${user.uid}/Expenses`)
             .snapshotChanges();
         } else {
-          return of([]); // Return an empty array as an observable when user is not logged in
+          return of([]);
         }
       })
     );
@@ -45,27 +46,50 @@ export class DataService {
     return this.fireauth.currentUser.then((user) => {
       if (user) {
         return this.afs
-          .doc(`/Users/${user.uid}/Expenses/${expense.id}`)
-          .delete();
+          .collection(`/Users/${user.uid}/Expenses`)
+          .ref.where('id', '==', expense.id)
+          .get()
+          .then((querySnapshot) => {
+            if (!querySnapshot.empty) {
+              const docId = querySnapshot.docs[0].id; // Get the actual Firestore document ID
+              return this.afs
+                .doc(`/Users/${user.uid}/Expenses/${docId}`)
+                .delete();
+            } else {
+              throw new Error('Expense not found');
+            }
+          });
       } else {
         throw new Error('User not logged in');
       }
     });
   }
 
-  updateExpense(expense: Expense) {
+  editExpense(expense: Expense) {
+    console.log('Update Expense Triggered');
     return this.fireauth.currentUser.then((user) => {
       if (user) {
         return this.afs
-          .doc(`/Users/${user.uid}/Expenses/${expense.id}`)
-          .update(expense);
+          .collection(`/Users/${user.uid}/Expenses`)
+          .ref.where('id', '==', expense.id)
+          .get()
+          .then((querySnapshot) => {
+            if (!querySnapshot.empty) {
+              const docId = querySnapshot.docs[0].id; // Get actual Firestore document ID
+              return this.afs
+                .doc(`/Users/${user.uid}/Expenses/${docId}`)
+                .update(expense);
+            } else {
+              throw new Error('Expense not found');
+            }
+          });
       } else {
         throw new Error('User not logged in');
       }
     });
   }
 
-  updateProfile(profile: Profile) {
+  createProfile(profile: Profile) {
     return this.fireauth.currentUser.then((user) => {
       if (user) {
         profile.id = this.afs.createId();
